# Dynamically add docstrings to client methods