@jamesls 2015-08-21
While the change looks good overall, I'm surprised that the 2-N times for creating a client is 3x slower.  I think we'll have to address this in the future.

Otherwise, :shipit: