@kyleknap 2015-08-21
Yeah that is something that we can look into. I am not too sure why it happened. Maybe some more iterations can be done to see if it is consistent.